+++ CHANGELOG.md
@@ -1,6 +1,8 @@
 # v0.6.0
 
 - Change how word wrapping is handled internally
+- Add some tests
+- Add more debug checks
 
 # v0.5.0